// webapp/src/shared/generated/models/Chatroom.ts
/* tslint:disable */
/* eslint-disable */
/**
 * transcendence-app
 * The transcendence-app API description
 *
 * The version of the OpenAPI document: 1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { exists, mapValues } from '../runtime';
/**
 * 
 * @export
 * @interface Chatroom
 */
export interface Chatroom {
    /**
     * 
     * @type {boolean}
     * @memberof Chatroom
     */
    isPublic: boolean;
    /**
     * 
     * @type {string}
     * @memberof Chatroom
     */
    id: string;
    /**
     * 
     * @type {string}
     * @memberof Chatroom
     */
    name: string;
    /**
     * 
     * @type {string}
     * @memberof Chatroom
     */
    avatarId: string;
    /**
     * 
     * @type {number}
     * @memberof Chatroom
     */
    avatarX: number;
    /**
     * 
     * @type {number}
     * @memberof Chatroom
     */
    avatarY: number;
    /**
     * 
     * @type {Date}
     * @memberof Chatroom
     */
    createdAt: Date;
    /**
     * 
     * @type {string}
     * @memberof Chatroom
     */
    ownerId: string;
}

/**
 * Check if a given object implements the Chatroom interface.
 */
export function instanceOfChatroom(value: object): boolean {
    let isInstance = true;
    isInstance = isInstance && "isPublic" in value;
    isInstance = isInstance && "id" in value;
    isInstance = isInstance && "name" in value;
    isInstance = isInstance && "avatarId" in value;
    isInstance = isInstance && "avatarX" in value;
    isInstance = isInstance && "avatarY" in value;
    isInstance = isInstance && "createdAt" in value;
    isInstance = isInstance && "ownerId" in value;

    return isInstance;
}

export function ChatroomFromJSON(json: any): Chatroom {
    return ChatroomFromJSONTyped(json, false);
}

export function ChatroomFromJSONTyped(json: any, ignoreDiscriminator: boolean): Chatroom {
    if ((json === undefined) || (json === null)) {
        return json;
    }
    return {
        
        'isPublic': json['isPublic'],
        'id': json['id'],
        'name': json['name'],
        'avatarId': json['avatarId'],
        'avatarX': json['avatarX'],
        'avatarY': json['avatarY'],
        'createdAt': (new Date(json['createdAt'])),
        'ownerId': json['ownerId'],
    };
}

export function ChatroomToJSON(value?: Chatroom | null): any {
    if (value === undefined) {
        return undefined;
    }
    if (value === null) {
        return null;
    }
    return {
        
        'isPublic': value.isPublic,
        'id': value.id,
        'name': value.name,
        'avatarId': value.avatarId,
        'avatarX': value.avatarX,
        'avatarY': value.avatarY,
        'createdAt': (value.createdAt.toISOString()),
        'ownerId': value.ownerId,
    };
}