// webapp/src/shared/generated/models/UpdateChatroomMemberDto.ts
/* tslint:disable */
/* eslint-disable */
/**
 * transcendence-app
 * The transcendence-app API description
 *
 * The version of the OpenAPI document: 1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { exists, mapValues } from '../runtime';
/**
 * 
 * @export
 * @interface UpdateChatroomMemberDto
 */
export interface UpdateChatroomMemberDto {
    /**
     * 
     * @type {boolean}
     * @memberof UpdateChatroomMemberDto
     */
    admin?: boolean;
    /**
     * 
     * @type {boolean}
     * @memberof UpdateChatroomMemberDto
     */
    muted?: boolean;
    /**
     * 
     * @type {boolean}
     * @memberof UpdateChatroomMemberDto
     */
    banned?: boolean;
}

/**
 * Check if a given object implements the UpdateChatroomMemberDto interface.
 */
export function instanceOfUpdateChatroomMemberDto(value: object): boolean {
    let isInstance = true;

    return isInstance;
}

export function UpdateChatroomMemberDtoFromJSON(json: any): UpdateChatroomMemberDto {
    return UpdateChatroomMemberDtoFromJSONTyped(json, false);
}

export function UpdateChatroomMemberDtoFromJSONTyped(json: any, ignoreDiscriminator: boolean): UpdateChatroomMemberDto {
    if ((json === undefined) || (json === null)) {
        return json;
    }
    return {
        
        'admin': !exists(json, 'admin') ? undefined : json['admin'],
        'muted': !exists(json, 'muted') ? undefined : json['muted'],
        'banned': !exists(json, 'banned') ? undefined : json['banned'],
    };
}

export function UpdateChatroomMemberDtoToJSON(value?: UpdateChatroomMemberDto | null): any {
    if (value === undefined) {
        return undefined;
    }
    if (value === null) {
        return null;
    }
    return {
        
        'admin': value.admin,
        'muted': value.muted,
        'banned': value.banned,
    };
}